+++ lib/oban/queue/executor.ex
@@ -109,9 +109,9 @@ defmodule Oban.Queue.Executor do
 
   @spec record_finished(t()) :: t()
   def record_finished(%__MODULE__{} = exec) do
-    stop_mono = System.monotonic_time(:microsecond)
+    stop_mono = System.monotonic_time()
     duration = stop_mono - exec.start_mono
-    queue_time = DateTime.diff(exec.job.attempted_at, exec.job.scheduled_at, :microsecond)
+    queue_time = DateTime.diff(exec.job.attempted_at, exec.job.scheduled_at, :native)
 
     %{exec | duration: duration, queue_time: queue_time, stop_mono: stop_mono}
   end